+++ changelog.md
@@ -1,5 +1,10 @@
 # ChangeLog
 
+## v2.4.3
+
+- 优化代码
+- Optimize the code
+
 ## v2.4.2
 
 - 区分静态锚文本和动态锚文本